[refactor] #1  reducer 함수 및 context 디스패치 함수 분리
